\name{internals}
\alias{internals}
\title{Internal Functions}
\description{
Functions for internal use only, or not yet documented.
}
\usage{
lpm.newton2(beta, XX, y, w, lambda, tol, maxit, safeit, gamma = 1e-5) 
lpm.fit(y, x, w, lambda, control)
lpm.newton(beta, X, y, w, lambda, tol, maxit, safeit)
L_lambda(beta, X, y, w, lambda, hessian = TRUE)
}
\value{No return value, called for side effect}
\keyword{internal}

